+++ api/src/main/java/it/unicam/cs/model/contenuti/Monumento.java
@@ -18,11 +18,11 @@ public final class Monumento extends POI {
     private int annoRealizzazione;
     private String descrizione;
     private String autore;
-    private int altezza;
-    private int lunghezza;
+    private double altezza;
+    private double lunghezza;
     private String architettura;
 
-    public Monumento(Integer id, String nome, Posizione posizione, Utente contributore, StatoElemento stato, Comune comuneAssociato, Indirizzo indirizzo, List<Evento> eventiAssociati, List<ContenutoMultimediale> contenutiMultimediali, int annoRealizzazione, String descrizione, String autore, int altezza, int lunghezza, String architettura) {
+    public Monumento(Integer id, String nome, Posizione posizione, Utente contributore, StatoElemento stato, Comune comuneAssociato, Indirizzo indirizzo, List<Evento> eventiAssociati, List<ContenutoMultimediale> contenutiMultimediali, int annoRealizzazione, String descrizione, String autore, double altezza, double lunghezza, String architettura) {
         super(id, nome, posizione, contributore, stato, comuneAssociato, indirizzo, eventiAssociati, contenutiMultimediali);
         this.annoRealizzazione = annoRealizzazione;
         this.descrizione = descrizione;
